# packages/dashboard/swagger_codegenerator/SwaggerDefintion.json
{"swagger":"2.0","info":{"description":"Geo fencer API description","version":"1.0","title":"Geo fencer"},"basePath":"/","tags":[{"name":"GeoFencer","description":""}],"schemes":["http"],"paths":{"/management/Rules":{"get":{"summary":"Get all rules from Geo Fencer service","description":"Get active GeoFencer rules","operationId":"GetRules","responses":{"200":{"description":"Returns the rules","schema":{"$ref":"#/definitions/RulesResult"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/AnalyseRule/{id}":{"get":{"summary":"Analyse a rule with all current states from simulation items","description":"Get current state of GeoFencer rule","operationId":"GetAnalyseRule","parameters":[{"type":"string","name":"id","required":true,"in":"path","description":"The rule ID."}],"responses":{"200":{"description":"Returns the current state of rule","schema":{"$ref":"#/definitions/AnalyseRuleResult"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/Definition":{"put":{"summary":"Upload geo fencer defintion file","description":"","operationId":"SetGeofencerDefinition","parameters":[{"name":"FakeGeoJSONEnvelopeInterface","required":true,"in":"body","schema":{"$ref":"#/definitions/FakeGeoJSONEnvelopeInterface"}}],"responses":{"200":{"description":"","schema":{"type":"string"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/SendSimulationTestData":{"put":{"summary":"Send simulation items to service (simulate kafka messages), ONLY FOR TESTING, WILL BE REMOVED IN FUTURE","description":"","operationId":"SendSimulationTestData","parameters":[{"name":"SimulationTestData","required":true,"in":"body","schema":{"$ref":"#/definitions/SimulationTestData"}}],"responses":{"200":{"description":"","schema":{"type":"string"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/DoTestRule":{"put":{"summary":"Test a rule evaluation (creating rules)","description":"Test rule","operationId":"DoTestRule","parameters":[{"name":"TestRule","required":true,"in":"body","schema":{"$ref":"#/definitions/TestRule"}}],"responses":{"200":{"description":"Outcome of rule test","schema":{"$ref":"#/definitions/ResultTestRule"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/Status":{"get":{"summary":"Get server status","description":"Get server status","operationId":"GetStatus","responses":{"200":{"description":"Returns the status","schema":{"$ref":"#/definitions/StatusResult"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"GeoFencerRule":{"type":"object","properties":{"UniqueSystemId":{"type":"string"},"UserAssignedId":{"type":"string"},"Rule":{"type":"string"},"IsRuleValid":{"type":"boolean"},"RuleError":{"type":"string"}},"required":["UniqueSystemId","UserAssignedId","Rule","IsRuleValid","RuleError"]},"RulesResult":{"type":"object","properties":{"Rules":{"type":"array","items":{"$ref":"#/definitions/GeoFencerRule"}}},"required":["Rules"]},"ItemState":{"type":"object","properties":{"ItemId":{"type":"string"},"Substituted":{"type":"string"},"InGeographicArea":{"type":"boolean"},"IsExpressionValid":{"type":"boolean"}},"required":["ItemId","Substituted","InGeographicArea","IsExpressionValid"]},"AnalyseRuleResult":{"type":"object","properties":{"Items":{"type":"array","items":{"$ref":"#/definitions/ItemState"}}},"required":["Items"]},"FakeGeoJSONEnvelopeInterface":{"type":"object","properties":{"Message":{"type":"string"}},"required":["Message"]},"SimulationTestData":{"type":"object","properties":{"JsonTextItemArray":{"type":"string"}},"required":["JsonTextItemArray"]},"TestRule":{"type":"object","properties":{"Expression":{"type":"string"},"Item":{"type":"string"}},"required":["Expression","Item"]},"ResultTestRule":{"type":"object","properties":{"Substituded":{"type":"string"},"ExpressionResult":{"type":"boolean"},"HasError":{"type":"boolean"},"ErrorMsg":{"type":"string"}},"required":["Substituded","ExpressionResult","HasError","ErrorMsg"]},"StatusResult":{"type":"object","properties":{"Description":{"type":"string"},"Version":{"type":"string"},"KafkaServer":{"type":"string"},"SchemaRegistryUrl":{"type":"string"}},"required":["Description","Version","KafkaServer","SchemaRegistryUrl"]}}}